requirements.txt
# Motor Glosas HUS — dependencias con versiones fijadas
# Actualizar periódicamente y correr pruebas tras cada actualización

fastapi==0.115.6
uvicorn[standard]==0.34.0
groq==0.13.1
PyPDF2==3.0.1
python-multipart==0.0.20
python-dotenv==1.0.1
jinja2==3.1.5
reportlab==4.2.5
pdfplumber==0.11.5
openpyxl==3.1.5
sqlalchemy==2.0.37
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.10
httpx>=0.27.0,<0.29.0
pydantic-settings>=2.2.0,<3.0.0
pydantic[email]>=2.0.0,<3.0.0
slowapi==0.1.9
cachetools==5.5.0
alembic==1.14.0
